Edits to correct filepaths as per meeting discussion
As was discussed - usage of a noeb directory (for non-easybuild built applications) is chosen retaining live and staging subdirectories if / when needed.

Structure is followed for modulefiles and appliation data and also follows the structure of ShARC's Ansys configuration using the binary file in a  subdirectory named after the ansys version.