This includes adding a reference info section so imports of specific sections can be done more than once.
Additionally this adds a global.rst file which is added at the front of all pages - this contains a substitution of |br| to make a <br/> in HTML inline where this is needed.

This is not done on each page in order to avoid multiple definition warnings. Further substitutions can be made as / when needed by adding these to the global.rst file.

Appropriate edits to the conf.py have been made to pull this in but also exclude it from being sourced.