Adding fluent reference info.
Requires re-organisation of the reference info file / folder structure and amendments to the toctrees to get the correct toctree display on the left working correct.

Fluent Reference info includes how to / why to use journal files as well as details on exporting plots/graphs in batch jobs with journal files.